grading system stuff

+++ grader.hpp
@@ -1,15 +1,14 @@
 #ifndef GRADER_HPP
 #define GRADER_HPP
-
 struct Card
 {
-    float grade;
+    double grade;
     int scratch_pixels;
     int dent_pixels;
     int fray_pixels;
 };
 
-float calculateGrade(Card &card);
+double calculateGrade(Card &card);
 void printResults(Card &card);
 
 #endif // GRADER_HPP
